tool_nhac/requirements.txt
# =======================================
# TikTok Audio Automation System V2.1
# Requirements — Full Stack & Web UI
# =======================================

# --- Web Framework (Real-time UI) ---
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-multipart>=0.0.9

# --- Web Automation ---
playwright>=1.42.0
playwright-stealth>=1.0.6

# --- HTTP & Async I/O ---
httpx>=0.27.0
aiofiles>=23.2.1

# --- Database ---
aiosqlite>=0.20.0

# --- AI / Audio Intelligence ---
shazamio>=0.4.0
faster-whisper>=0.10.1
pydub>=0.25.1
imageio-ffmpeg>=0.4.9
librosa>=0.10.1
soundfile>=0.12.1
google-generativeai>=0.4.1

# --- Machine Learning / Ranking ---
# (Duoc cai dat thong qua setup.bat de toi uu GPU/CPU)
sentence-transformers>=2.7.0

scikit-learn>=1.4.0
numpy>=1.26.0

# --- V2.1: VAD (Voice Activity Detection) ---
# Silero VAD — siêu nhẹ, chạy trên CPU, phát hiện giọng người trong 0.05s
# Load qua torch.hub (không cần cài riêng, torch.hub.load sẽ tự download)
# Nếu muốn cài offline: pip install silero-vad
# silero-vad>=5.1.0     # Optional — torch.hub.load tự xử lý

# --- V2.1: YAMNet Music Classifier (Optional) ---
# Cài nếu muốn dùng YAMNet để phát hiện nhạc nền lofi/no-copyright
# Chỉ chạy trong vùng xám (speech_ratio 0.40–0.65), giảm tải GPU đáng kể
# tensorflow>=2.15.0    # Optional — comment out nếu không dùng YAMNet
# tensorflow-hub>=0.16.0
# resampy>=0.4.3        # Optional — dùng để resample audio cho YAMNet

# --- Data Handling ---
pandas>=2.2.0
openpyxl>=3.1.2        # Đọc file Excel (.xlsx) cho channel_expander.py

# --- Utilities ---
python-dotenv>=1.0.1
tenacity>=8.2.3
loguru>=0.7.2
rich>=13.7.0
pydantic>=2.6.0
